Remove usage of `asynccontextmanager` which is not available for all Python versions that `proxy.py` supports
